+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2026-01-01
+
+### Changed
+- Time is now prepended at the beginning of the message instead of appended at the end
+- Added deduplication to prevent time appearing twice if hook is called multiple times
+
+### Fixed
+- Fixed duplicate time stamps appearing in messages
+
 ## [0.6.0] - 2026-01-01
 
 ### Changed